changed the cast from int to long long and it seems to fix the hang for set_differece for large indices

+++ thrust/system/hip/detail/set_operations.h
@@ -61,7 +61,7 @@ namespace __set_operations
         T    key2 = data[mid];
         bool pred = UpperBound ? !comp(key, key2) : comp(key2, key);
         if(pred)
-            begin = (int)mid + 1;
+            begin = (long long)mid + 1;
         else
             end = mid;
     }
